add: base board controller
